Promote local tempdir context manager to 'g.c._testing'.
